refactor: separate SwiperList, PetCard and NoPet from overview container
